Added main portal and portal piece item
