+++ Assets/_Environment/NonDistructables/Materials.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 16655e88059fc714fb4dcc1d9a86649e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
