+++ Assets/_Environment/NonDistructables/Materials/Main_Portal.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e5dc4e79e1d935847aaba5b518f92acf
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
