+++ Assets/_Environment/NonDistructables/Prefabs/Main Portal.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: ae514810145423f4d9cc84d7b2a66784
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
