+++ Assets/_Environment/NonDistructables/Scripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1ce0107a77e5249499cf9acea4eaa79f
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
