+++ Assets/_Environment/NonDistructables/Scripts/PortalInteraction.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a4975f4b77ba5e1499ebe2028fe627e1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
