+++ Assets/_GameController/Prefabs/GameController.prefab
@@ -23797,6 +23797,7 @@ MonoBehaviour:
   - {fileID: 586135968040543494, guid: 5772fba190d095a4b9fb3fd87cce3ca2, type: 3}
   - {fileID: 567451159922232694, guid: fa2c4cdefa2cea44ab42f5a49eafd9cc, type: 3}
   - {fileID: 1442854967829722669, guid: 804573bf0314ea1488557092258ba243, type: 3}
+  - {fileID: 2014520146305032985, guid: d910c247c60a0c946ab5753c646b611b, type: 3}
   environmentItemList:
   - {fileID: 1905376070424928, guid: b16bdde1c27ac4d46b0afd1342c646bf, type: 3}
   - {fileID: 5117017173607214059, guid: 221a0c3dadfb9dd489f5b8809ad35baa, type: 3}
