+++ Assets/_Items/General/Mesh/PortalFragment.mtl
@@ -1,0 +1,5 @@
+newmtl Main_Portal
+bump bake_plane_tri_normal_map_from_mesh.png
+Kd 0.3119883 0.3481006 0.3867925
+d 1
+
