+++ Assets/_Items/General/Mesh/PortalFragment.mtl.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 7cfdb2d31e7662049857af131aa5887d
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
