+++ Assets/_Items/General/Mesh/PortalFragment.obj
@@ -1,0 +1,53 @@
+# ProBuilder 5.0.6
+# https://unity3d.com/unity/features/worldbuilding/probuilder
+# 12/29/2023 1:19:57 PM
+
+mtllib ./PortalFragment.mtl
+o PortalFragment
+
+g PortalFragment
+v 1.5 0 0.29239
+v -1.5 0 0.29239
+v 0.9173365 1.401664 0.2356017
+v -0.9173365 1.401664 0.2356017
+v -1.5 0 0
+v -0.9173365 1.401664 0.05678833
+v 1.5 0 0
+v 0.9173365 1.401664 0.05678833
+
+vt 0 -0.3838499
+vt -1 -0.3838499
+vt -0.1942212 0.1221321
+vt -0.8057789 0.1221321
+vt 1 -0.3838499
+vt 0.8057789 0.1221321
+vt 0.1942212 0.1221321
+vt 1 0
+vt 0 0
+vt 0.8057789 0.505982
+vt 0.1942212 0.505982
+vt -1 0
+vt -0.1942212 0.505982
+vt -0.8057789 0.505982
+vt 0.1942212 0.8057789
+vt 0.8057789 0.8057789
+vt 0.1942212 0.1942212
+vt 0.8057789 0.1942212
+vt 0 1
+vt -1 1
+
+vn 0 0.0404817 0.9991803
+vn -0.9233955 0.38385 0
+vn 0 0.04048171 -0.9991803
+vn 0.9233955 0.38385 0
+vn 0 1 0
+vn 0 -1 0
+
+usemtl Main_Portal
+f 3/3/1 4/4/1 2/2/1 1/1/1
+f 4/6/2 6/7/2 5/1/2 2/5/2
+f 6/10/3 8/11/3 7/9/3 5/8/3
+f 8/13/4 3/14/4 1/12/4 7/9/4
+f 8/17/5 6/18/5 4/16/5 3/15/5
+f 1/19/6 2/20/6 5/12/6 7/9/6
+
