+++ Assets/_Items/General/Prefabs/PortalFragment.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: d910c247c60a0c946ab5753c646b611b
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
